+++ R/zzz.R
@@ -1,7 +1,7 @@
 pkg <- new.env(parent = emptyenv())
 
 .onLoad <- function(libname, pkgname) {
-  pkg$catalog_url <- "https://data.bedrock.bio/catalog.json"
+  pkg$catalog_url <- "https://data.bedrock.bio/manifest.json"
   pkg$credentials_url <- "https://data.bedrock.bio/credentials.json"
 }